requirements.txt
fastapi[standard]
python-dotenv
requests
sqlalchemy
# pyodbc
python-multipart
# bcrypt
# passlib
pymongo
redis
